+++ WorkspaceServer/Kernel/CSharpRepl.cs
@@ -13,7 +13,6 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Scripting;
 using Microsoft.CodeAnalysis.Scripting;
-using Microsoft.CodeAnalysis.VisualBasic.Syntax;
 
 namespace WorkspaceServer.Kernel
 {
@@ -153,35 +152,6 @@ namespace WorkspaceServer.Kernel
                 default:
                     throw new KernelCommandNotSupportedException(command, this);
             }
-        }
-    }
-
-    internal static class EnumerableExtensions
-    {
-        public static IOrderedEnumerable<T> OrderBy<T>(this IEnumerable<T> source, Comparison<T> compare)
-        {
-            var comparer = Comparer<T>.Create(compare);
-            return source.OrderBy(t => t, comparer);
-        }
-    }
-
-    internal static class ScriptExtensions
-    {
-        public static IEnumerable<Diagnostic> GetDiagnostics(this Script script)
-        {
-            if(script == null)
-            {
-                return Enumerable.Empty<Diagnostic>();
-            }
-
-            var compilation = script.GetCompilation();
-            var orderedDiagnostics = compilation.GetDiagnostics().OrderBy((d1, d2) =>
-            {
-                var severityDiff = (int)d2.Severity - (int)d1.Severity;
-                return severityDiff != 0 ? severityDiff : d1.Location.SourceSpan.Start - d2.Location.SourceSpan.Start;
-            });
-
-            return orderedDiagnostics;
         }
     }
 }
