+++ WorkspaceServer/Kernel/EnumerableExtensions.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace WorkspaceServer.Kernel
+{
+    internal static class EnumerableExtensions
+    {
+        public static IOrderedEnumerable<T> OrderBy<T>(this IEnumerable<T> source, Comparison<T> compare)
+        {
+            var comparer = Comparer<T>.Create(compare);
+            return source.OrderBy(t => t, comparer);
+        }
+    }
+}
